refactor: move metadata to separated feature
