+++ frontend/tsconfig.node.json
@@ -24,12 +24,15 @@
       "@/*": ["./src/*"],
       "@/components/*": ["./src/shared/components/*"],
       "@components/*": ["./src/shared/components/*"],
+      
       "@student/*": ["./src/features/student/*"],
       "@program/*": ["./src/features/program/*"],
       "@course/*": ["./src/features/course/*"],
       "@faculty/*": ["./src/features/faculty/*"],
       "@status/*": ["./src/features/status/*"],
       "@settings/*": ["./src/features/settings/*"],
+      "@metadata/*": ["./src/features/metadata/*"],
+
       "@ui/*": ["./src/shared/components/ui/*"],
       "@core/*": ["./src/core/*"],
       "@features/*": ["./src/features/*"],
